homepairs/homepairsUI/unittests/requirements.txt
altgraph==0.16.1
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0
djangorestframework==3.10.3
enum34==1.1.6
future==0.18.2
gunicorn==20.0.4
idna==2.8
ipaddress==1.0.23
kiwisolver==1.1.0
macholib==1.11
matplotlib==3.1.2
modulegraph==0.17
more-itertools==8.0.0
numpy==1.17.4
psycopg2-binary==2.8.4
py2app==0.19
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.8
whitenoise==4.1.4
xattr==0.9.7
